backend-server/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
sqlalchemy==2.0.27
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
email-validator==2.1.0.post1
fastapi-admin==1.0.0
aioredis==2.0.1
python-binance==1.0.19
web3==6.15.1
aiohttp==3.9.3 